added menu bar to all pages
