+++ src/AboutPage.jsx
@@ -1,0 +1,15 @@
+import React, { Component } from "react";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import NavBar from "./Components/NavBar/Navbar";
+
+function AboutPage() {
+  return (
+    <div>
+      <div>
+        <h1>About Us</h1>
+      </div>
+    </div>
+  );
+}
+
+export default AboutPage;
