--- src/Components/DetailsPage.jsx
@@ -1,12 +1,0 @@
-import React, { Component } from "react";
-import { BrowserRouter, Route, Switch } from "react-router-dom";
-
-function DetailsPage() {
-  return (
-    <div className="MyHome">
-      <p>Details Page Test</p>
-    </div>
-  );
-}
-
-export default DetailsPage;
