+++ src/Components/NavBar/SidebarData.js
@@ -1,12 +1,11 @@
 import React from "react";
 import * as FaIcons from "react-icons/fa";
 import * as AiIcons from "react-icons/ai";
-import * as IoIcons from "react-icons/io";
 
 export const SidebarData = [
   {
     title: "Explore",
-    path: "/",
+    path: "/explore",
     icon: <AiIcons.AiOutlineSearch />,
     cName: "nav-text",
   },
@@ -14,6 +13,12 @@ export const SidebarData = [
     title: "Results",
     path: "/results",
     icon: <AiIcons.AiOutlineUnorderedList />,
+    cName: "nav-text",
+  },
+  {
+    title: "Details",
+    path: "/details",
+    icon: <FaIcons.FaCookieBite />,
     cName: "nav-text",
   },
   {
