--- src/Components/ResultsPage.jsx
@@ -1,12 +1,0 @@
-import React, { Component } from "react";
-import { BrowserRouter, Route, Switch } from "react-router-dom";
-
-function ResultsPage() {
-  return (
-    <div className="MyHome">
-      <p>Results Page Test</p>
-    </div>
-  );
-}
-
-export default ResultsPage;
