+++ src/DetailsPage.jsx
@@ -1,0 +1,15 @@
+import React, { Component } from "react";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import NavBar from "./Components/NavBar/Navbar";
+
+function DetailsPage() {
+  return (
+    <div>
+      <div>
+        <h1>Details Page</h1>
+      </div>
+    </div>
+  );
+}
+
+export default DetailsPage;
