+++ src/ExplorePage.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+
+function ExplorePage() {
+  return (
+    <div>
+      <div>
+        <h1>Explore Page</h1>
+      </div>
+    </div>
+  );
+}
+
+export default ExplorePage;
