+++ src/ResultsPage.jsx
@@ -1,0 +1,15 @@
+import React, { Component } from "react";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import NavBar from "./Components/NavBar/Navbar";
+
+function ResultsPage() {
+  return (
+    <div>
+      <div>
+        <h1>Results Page</h1>
+      </div>
+    </div>
+  );
+}
+
+export default ResultsPage;
